version 0.59p by Michael Hipp:
- Added mpglib (see mpglib directory)
- 'make generic' is more generic now
- Added HPUX/ALib support from Ducroquet Erwan (ducroque@ufr-info-p7.ibp.fr) 
- Added EsoundD support from <fill it in>
- Added exp. relatime support (compile with -DSET_RT) 
- Added linux-sparc patch from <hmm lost email)
- Added genrepatch from Shane Wegner <shane@CM.NU>
- Added wav-File patch Samuel Audet <guardia@cam.org>
- Added SGI-RAD support from "J. Dean Brederson" <jdb@cs.utah.edu>
- Added sajberplay/FreeBSD patch from Chou Ye-chi <is84002@cis.nctu.edu.tw>
- Added 486 optimizations from Fabrice Bellard <bellard@email.enst.fr>
- Added combined output devices on SUN (ie Speaker and headphones together)
- Added audio_alsa.c (under GPL license) for the ALSA sound system from
   A. Hermansen <ahermans@vf.telia.no> and J. Kysela <perex@jcu.cz>
-Added K6-3DNow patches from  KIMURA Takuhiro <kim@hannah.ipc.miyakyo-u.ac.jp>



git-svn-id: svn://scm.orgis.org/mpg123/mhipp@31 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ getlopt.c
@@ -48,7 +48,7 @@ int performoption (int argc, char *argv[], topt *opt)
 			if (opt->flags & 2) /* var is *char */
 				*((char *) opt->var) = (char) opt->value;
 			else
-				*((int *) opt->var) = opt->value;
+				*((long *) opt->var) = opt->value;
 		else
 			result = opt->value ? opt->value : opt->sname;
 	else { /* requires argument */
@@ -60,7 +60,7 @@ int performoption (int argc, char *argv[], topt *opt)
 			if (opt->flags & 2) /* var is *char */
 				*((char **) opt->var) = strdup(loptarg);
 			else
-				*((int *) opt->var) = atoi(loptarg);
+				*((long *) opt->var) = atoi(loptarg);
 		else
 			result = opt->value ? opt->value : opt->sname;
 	}
